refactor: extract api/authentication.ts and api/questions.ts
