+++ src/api/questions.ts
@@ -1,0 +1,5 @@
+import "firebase/firestore"
+
+import firebase from "./firebase"
+
+export const questionsRef = firebase.firestore().collection("questions")
